# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Text editor MVP API
  version: 1.0.0
  description: Identity, Document and Processing contracts; ownership, folder and
    save semantics in docs.
servers:
  - url: http://localhost:8080
    description: Local same-origin reverse proxy
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/documents:
    get:
      tags:
        - Document
      x-service: document-service
      operationId: listDocuments
      summary: List owned/shared/trash documents
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentPage'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      description: parent filters OWNED only; SHARED ignores folder scope and must
        not disclose folderId (returns null). TRASH is owner only. Title search is
        metadata search, not full text index.
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum:
              - OWNED
              - SHARED
              - TRASH
            default: OWNED
        - *id002
        - name: titleQuery
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
        - *id003
        - *id004
    post:
      tags:
        - Document
      x-service: document-service
      operationId: createDocument
      summary: Create empty head revision zero
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
  /api/v1/documents/{documentId}:
    get:
      tags:
        - Document
      x-service: document-service
      operationId: getDocument
      summary: Metadata plus effective role
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      description: Non-owner responses redact folderId to null. Active documents only;
        trash metadata is in the owner TRASH list.
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
    patch:
      tags:
        - Document
      x-service: document-service
      operationId: patchDocument
      summary: Owner rename or move
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      description: At least one of title,folderId required; explicit null moves to
        root. Does not change content headRevision.
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchDocumentRequest'
    delete:
      tags:
        - Document
      x-service: document-service
      operationId: trashDocument
      summary: Owner move to trash
      responses:
        '204':
          description: Success with no body.
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevisionMutationRequest'
  /api/v1/documents/{documentId}/restore:
    post:
      tags:
        - Document
      x-service: document-service
      operationId: restoreDocument
      summary: Owner restore from trash
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevisionMutationRequest'
  /api/v1/documents/{documentId}/versions:
    get:
      tags:
        - Document
      x-service: document-service
      operationId: listVersions
      summary: List retained immutable versions
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionPage'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
        - *id003
        - *id004
    post:
      tags:
        - Document
      x-service: document-service
      operationId: commitVersion
      summary: Validate ticket and atomically commit immutable version
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
        '200':
          description: Valid no-change snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
        - name: Idempotency-Key
          in: header
          required: true
          schema: *id001
          description: Stable UUID for the exact request body; retry with the same key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitRequest'
  /api/v1/documents/{documentId}/versions/{revision}/download:
    get:
      tags:
        - Document
      x-service: document-service
      operationId: getVersionDownload
      summary: Authorize exact snapshot download
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadDescriptor'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
        '410':
          $ref: '#/components/responses/ErrorResponse'
      description: Touch last_access_at under version lock. Returns immutable objectRef
        to an authorized caller and a URL valid <=60s. Retired revision returns410
        VERSION_GONE. No endpoint for revision0; client uses empty model.
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
        - name: revision
          in: path
          required: true
          schema: *id005
  /api/v1/documents/{documentId}/uploads:
    post:
      tags:
        - Document
      x-service: document-service
      operationId: createUpload
      summary: Create bounded upload ticket
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadTicket'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadRequest'
  /api/v1/uploads/{uploadId}/content:
    put:
      tags:
        - Document
      x-service: document-service
      operationId: uploadLocalContent
      summary: Stream bytes to local storage adapter
      responses:
        '204':
          description: Success with no body.
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      description: LOCAL adapter only; actor and current EDITOR/OWNER grant checked.
        Cloud browser uploads to the GCS session URI instead. Create-new file, no
        overwrite.
      parameters:
        - name: uploadId
          in: path
          required: true
          schema: *id001
        - name: Content-Length
          in: header
          required: true
          schema:
            type: integer
            minimum: 20
            maximum: 33554432
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
...
  /api/v1/documents/{documentId}/versions/{revision}/content:
    get:
      tags:
        - Document
      x-service: document-service
      operationId: streamVersionContent
      summary: Authenticated native stream for LOCAL or proxy fallback
      responses:
        '200':
          description: Native snapshot bytes; no public storage URL.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      description: Same ACL and retention check as download descriptor. Exact immutable
        generation; no-store. Used when signed URL is unavailable.
      parameters:
        - name: documentId
          in: path
          required: true
          schema: *id001
        - name: revision
          in: path
          required: true
          schema: *id005
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          maxLength: 64
        message:
          type: string
          maxLength: 500
        traceId:
          type: string
          maxLength: 64
        details:
          type: object
          additionalProperties: true
      additionalProperties: false
      required:
        - code
        - message
        - traceId
...
    Document:
      type: object
      properties:
        id: *id001
        ownerUserId: *id001
        folderId: *id007
        title:
          type: string
          maxLength: 200
        headRevision: &id008
          type: integer
          format: int64
          minimum: 0
          maximum: 9007199254740991
        headVersionId: *id007
        metadataRevision: *id005
        effectiveRole:
          type: string
          enum:
            - OWNER
            - EDITOR
            - VIEWER
        createdAt: *id006
        updatedAt: *id006
        deletedAt:
          type: string
          format: date-time
          nullable: true
        preview:
          type: object
          nullable: true
          additionalProperties: true
      additionalProperties: false
      required:
        - id
        - ownerUserId
        - folderId
        - title
        - headRevision
        - headVersionId
        - metadataRevision
        - effectiveRole
        - createdAt
        - updatedAt
        - deletedAt
    CreateDocumentRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        folderId: *id007
      additionalProperties: false
      required:
        - title
        - folderId
    PatchDocumentRequest:
      type: object
      properties:
        expectedMetadataRevision: *id005
        title:
          type: string
          minLength: 1
          maxLength: 200
        folderId: *id007
      additionalProperties: false
      required:
        - expectedMetadataRevision
      minProperties: 2
    RevisionMutationRequest:
      type: object
      properties:
        expectedMetadataRevision: *id005
      additionalProperties: false
      required:
        - expectedMetadataRevision
    ObjectRef:
      type: object
      properties:
        provider:
          type: string
          enum:
            - LOCAL
            - GCS
        bucket:
          type: string
          nullable: true
          maxLength: 128
        key:
          type: string
          maxLength: 512
        generation:
          type: string
          pattern: ^[0-9]+$
          maxLength: 32
      additionalProperties: false
      required:
        - provider
        - bucket
        - key
        - generation
    Version:
      type: object
      properties:
        id: *id001
        documentId: *id001
        revision: *id005
        nativeSha256: &id009
          type: string
          pattern: ^[a-f0-9]{64}$
        nativeBytes:
          type: integer
          minimum: 20
          maximum: 33554432
        textUtf8Bytes:
          type: integer
          minimum: 0
          maximum: 10485760
        utf16Length:
          type: integer
          minimum: 0
          maximum: 10485760
        logicalLines:
          type: integer
          minimum: 1
          maximum: 1000000
        createdByUserId: *id001
        createdAt: *id006
      additionalProperties: false
      required:
        - id
        - documentId
        - revision
        - nativeSha256
        - nativeBytes
        - textUtf8Bytes
        - utf16Length
        - logicalLines
        - createdByUserId
        - createdAt
    UploadRequest:
      type: object
      properties:
        expectedHeadRevision: *id008
        nativeBytes:
          type: integer
          minimum: 20
          maximum: 33554432
        nativeSha256: *id009
      additionalProperties: false
      required:
        - expectedHeadRevision
        - nativeBytes
        - nativeSha256
    UploadTicket:
      type: object
      properties:
        uploadId: *id001
        kind:
          type: string
          enum:
            - LOCAL
            - GCS_RESUMABLE
        uploadUrl:
          type: string
          format: uri
        method:
          type: string
          enum:
            - PUT
        requiredHeaders:
          type: object
          additionalProperties:
            type: string
        commitDeadlineAt: *id006
      additionalProperties: false
      required:
        - uploadId
        - kind
        - uploadUrl
        - method
        - requiredHeaders
        - commitDeadlineAt
    CommitRequest:
      type: object
      properties:
        uploadId: *id001
        expectedHeadRevision: *id008
      additionalProperties: false
      required:
        - uploadId
        - expectedHeadRevision
    SaveResponse:
      type: object
      properties:
        version:
          $ref: '#/components/schemas/Version'
        noChange:
          type: boolean
      additionalProperties: false
      required:
        - version
        - noChange
    DownloadDescriptor:
      type: object
      properties:
        url:
          type: string
          format: uri
        expiresAt: *id006
        version:
          $ref: '#/components/schemas/Version'
        objectRef:
          $ref: '#/components/schemas/ObjectRef'
      additionalProperties: false
      required:
        - url
        - expiresAt
        - version
        - objectRef
...
    DocumentPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        nextCursor:
          type: string
          nullable: true
      additionalProperties: false
      required:
        - items
        - nextCursor
    VersionPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Version'
        nextCursor:
          type: string
          nullable: true
      additionalProperties: false
      required:
        - items
        - nextCursor
...
  responses:
    ErrorResponse:
      description: Error with stable code; semantics and status in docs/05_SAVE_AND_SHARE.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'